Add Identifier class
This class abstracts over identifiers, allowing us to lock down the API
and logic of inspecting and comparing identifiers that may exist within
a path. The idea is, given an identifier like `foo::bar::baz`, we may
want to get the final name, `baz`, or the path to the parent, `foo::bar`, or
just the root of the path `foo`, and we may wish to compare two paths
for equality and construct new paths, etc. But having to implement all
of this logic using string munging or array manipulation directly is
error prone, so we should protect the operations in an ADT.