Switch WithTypeAnnotation for WithType in IR
The `WithType` and `WithOptionalType` interfaces were added with the
expectation that they would hold the type information for the enclosing
term in the AST. However, as @bugarela explained to me yesterday, we are
actually storing the typing information in a separate map, and
coordinating those with terms using their type id.

There is however a use for this fields currently, which is to store
type annotations.

This change set:

- reflects the field's current use by calling them `WithTypeAnnotation`
  and `WithOptionalTypeAnnotation`.
- drop these fields from all structures which cannot be annotated