Singleton operator type params can omit parens
Closes #455

The language spec indicates that we support operator type annotations
without parens around the params when there is only a single argument.
This is also consistent with our syntax for lambdas, which allow `a =>
b` as well as `(a) => b`.

This small change adjusts the grammar rules to match the spec and be
more consistent with the syntax for our values.

It has the drawback of making two parsing errors in tests slightly
worse, but suggesting an operator type might appear where it shouldn't.
However, this is consistent with our current grammar rules, and should
be fixed by https://github.com/informalsystems/tnt/issues/456 once we
decide on the course of action there.