Make entrypoint executable
The compilation to js was not producing a `cli.js` with executable
permissions. As a result, installations on my machine using nvm
sandboxes were broken with a cryptic error:

```
$  quint
bash: ~/.nvm/versions/node/v19.6.0/bin/quint: Permission denied
```

Not sure why this started showing up now for me, but ensuring we make
the entrypoint executable fixes it for me.