Remove redundant type constraints on record label strings
The way we construct and deconstruct records is interesting: in the
normal form we interpret a special set of operators such that
string-valued expressions given as arguments in our object language are
"lifted" into product type indexes in our meta language. E.g., in `label(r,
"foo")` we extract the value of the internal representation of the quint
string `"foo"` and use this as the label for projecting a value out of
the record `r`.

This lets us do some interesting (and perhaps undesirable?) things, like
construct labels which cannot be used to project out values using the
`r._` syntax sugar:

```
>>> Rec(":", 1)
{ :: 1 }
```

In the process of lifting the value of these strings into labels, we
also verify that each label supplied is in fact a string letteral:

E.g.,

https://github.com/informalsystems/quint/blob/be8bb62dbc046da134d07f44367c4a60970adb45/quint/src/types/specialConstraints.ts#L93-L99

We were previously *also* adding constraints to the our type checking
that reasserted this invariant, but those constraints aren't necessary.
For the core reason that the operator that construct, update, and
eliminate records are not "normal operators" but special constructs, and
the strings (or, for tuples) supplied to describe the labels do not end
up as expressions that require type constraints, but are "puns" that
represent the index in question.