Fix type declarations in records.qnt example
Free type variables should not occur in variable or constant
declarations, as these are required to be concrete, monomorphic values
in our current language design.

The underlying issue that these annotations are allowed by the type
checker is reported at
https://github.com/informalsystems/quint/issues/789#issuecomment-1648279005

Fixing the type annotations resolves the type inference problem that was
tripping up apalache.