Support multiple declarations in the REPL
This is required in order to support sum type declarations in the REPL.

Sum type declarations work like this this: When a type alias for a sum
is declared, e.g., `type T = A | B`, the parser expands this into 3 declarations:

```
type T = A | B

val A : T = variant("A", {})
val B : T = variant("B", {})
```

However, the previous implementation the read and eval parts of the REPL
took the input read, ran it thru the parser, then took only the *last*
declaration, ignoring anything else. As a result, the type annotation
sealing the generated operators with the proper sum type was in the
context during type checking.

This behavior of dropping all but the last declaration is completely
unnecessary, so this change generalizes the read and eval to work on any
sequence of declarations that are parsed out of the input.