Remove GH bear token from rust backend download
GitHub seems to validate tokens regardless of the resource visibility.
If the user has an invalid token in their environment, downloading the
rust backend will fail even though it's publicly available in the
project's release assets.

This patch removes the GITHUB_TOKEN from the requests to list and
download assets from GitHub to avoid unauthorized issues. Moreover, it
includes an integration test that verifies quint is able to download the
latest rust banckend for further coverages of this workflow.