Normalize Apalache traces
Apalache's traces are not normalized like quint simulator traces are.
That means, for example, that maps and sets are going to be printed with
elements in random order as opposed to alphabetically sorted.

This patch makes use of quint runtime values conversions to normalize
Apalache produced traces for better readability and conformance with
quint generated traces.