Print commands in file's expression history
When entering the repl, if expressions exist in the form of /*! expr !*/
comments in the quint source code, those are evaluated by the repl upon
loading the file. However, the repl won't show what commands were
executed, only their output.

This patch changes this behavior so that when loading a module, unless
in quiet mode, the reply will show which commands were executed after
loading.