Move rust integration tests to a pre-release step
Rust integration tests are flaky when running on a per PR basis due to
the GH API rate limit that prevents us from downloading the evaluator
multiple times in parallel. This patch moves the integration test to a
pre-release step so it runs only as a sanity check before releases.