Fix double free when using LocalHipStr
The LocalHipStr is allowed to cross thread boundaries by implementing
the Send auto trait. However, on large heap-allocated strings, it causes
a double free due to unsynchronized drops.

For the moment, I'm changing the representation of strings to the
thread-safe version of HipStr since I've tested it and it works well.